Lossless decoder fix for a special transform order
Fix the lossless decoder for the case when it has to apply other inverse
transforms before applying Color indexing inverse transform.

The main idea is to make ColorIndexingInverse virtually in-place: we use
the fact that the argb_cache is allocated to accommodate all *unpacked*
pixels of a macro-row, not just *packed* pixels.

Change-Id: I27f11f3043f863dfd753cc2580bc5b36376800c4