add alpha dithering for lossy
new options:
 dwebp -alpha_dither
 vwebp -noalphadither

When the source was marked as quantized, we use a threshold-averaging
filter to smooth the decoded alpha plane.
Note: this option forces the decoding of alpha data in one pass, and
might slow the decoding a bit.

The new flag in WebPDecoderOptions struct is 'allow_alpha_dithering'
(off by default)

Change-Id: I218e21af96360d4781587fede95f8ea4e2b7287a