QuantizeBlock SSE2 Optimization:
Store to Load forward block was detected in the function QuantizeBlock.
The store was done in the function ReconstructIntra16 the first
element in the two dimensional tmp array was zeroed before being sent to
QuantizeBlock function.
The solution to that was zeroing the first element after the 16 elements in
the tmp array are being loaded
This fix gives ~10% function level gain and ~2% user level gain

Change-Id: Iad228468bf3cfcbc5400b13adedac2043fab18eb