PickBestIntra4 for AVX2 - Part II
The inner loop of PickBestIntra4 can now handle two iterations at a time
using AVX2.
Every iteration used to process 4 bytes and now it process 8 bytes.
This commit is the second part of "PickBestIntra4" for AVX2 which include
the main function "PickBestIntra4" and its subfunctions that call AVX2
functions.
There is a 24% function level gain and 12% user level gain.

Change-Id: Ic41c3c68f2964d81d8f85ddbee572986066ea1f2