encoder: switch BPS to 32 instead of 16
this is a first step to unifying encoding/decoding cache stride
and possibly sharing the prediction functions in dsp/

With this layout, there's a little (~7%) space lost with unused samples.
But no speed change was observed.

WARNING! I had to disable MIPS32 code for now, since BPS is
too deeply hardcoded.

Change-Id: I016df8cad41bde5088df3579e6ad65d884ee711e